Set windmove functions in the global keymap.
Windmove has the "windmove-default-keybindings" function to set the default
keybindings and this worked in the past. However, it seems to add the
keybindings to a local keymap now, which ends up taking predecence even in
org-mode buffers. By setting the keybindings manually in the global keymap we
allow org-mode to easily replace them in org-mode buffers (and call windmove
only when appropriately).